requirements.txt
# requirements.txt

torch>=1.10
torchvision>=0.11
numpy
opencv-python
scikit-learn
tqdm
Pillow
facenet-pytorch
timm